requirements.txt
fastapi~=0.111.0
uvicorn
httpx
python-dotenv~=1.0.1

openai~=0.27.8
pydantic~=2.7.1